+++ packages/core/test/MiddlewareSpec.test.ts
@@ -38,20 +38,20 @@ class MutationOnly extends MiddlewareSpec.MiddlewareSpec<MutationOnly>()(
 
 const query = FunctionSpec.publicQuery({
   name: "getThing",
-  args: () => Schema.Struct({}),
+  args: () => ({}),
   returns: () => Schema.String,
 });
 
 const queryWithError = FunctionSpec.publicQuery({
   name: "getThingOrFail",
-  args: () => Schema.Struct({}),
+  args: () => ({}),
   returns: () => Schema.String,
   error: () => NotFound,
 });
 
 const mutation = FunctionSpec.publicMutation({
   name: "setThing",
-  args: () => Schema.Struct({}),
+  args: () => ({}),
   returns: () => Schema.Null,
 });