+++ packages/server/test/CronJob.test.ts
@@ -10,7 +10,7 @@ const makeMutationRef = (functionNamespace: string, name: string) =>
     functionNamespace,
     FunctionSpec.internalMutation({
       name,
-      args: () => Schema.Struct({}),
+      args: () => ({}),
       returns: () => Schema.Void,
     }),
   );
@@ -20,7 +20,7 @@ const makeMutationRefWithArgs = (functionNamespace: string, name: string) =>
     functionNamespace,
     FunctionSpec.internalMutation({
       name,
-      args: () => Schema.Struct({ email: Schema.String }),
+      args: () => ({ email: Schema.String }),
       returns: () => Schema.Void,
     }),
   );